# Cache access to namespace elements to improve performance